Add serialization support for big arrays
This is gated on the const-generics support available in version 1.51+.
This includes basic nesting of types with serde_as and the conversion to a map.
