+++ Cargo.toml
@@ -36,6 +36,7 @@ macros = ["serde_with_macros"]
 chrono_crate = {package = "chrono", version = "0.4.1", features = ["serde"], optional = true}
 doc-comment = {version = "0.3.3", optional = true}
 hex = {version = "0.4.2", optional = true}
+rustversion = "1.0.0"
 serde = {version = "1.0.75", features = ["derive"]}
 serde_json = {version = "1.0.1", optional = true}
 serde_with_macros = {path = "./serde_with_macros", version = "1.4.1", optional = true}
@@ -48,7 +49,6 @@ mime = "0.3.16"
 pretty_assertions = "0.6.1"
 regex = {version = "1.3.9", default-features = false, features = ["std"]}
 ron = "0.6"
-rustversion = "1.0.0"
 serde-xml-rs = "0.4.1"
 serde_derive = "1.0.75"
 serde_json = {version = "1.0.25", features = ["preserve_order"]}
